Moving generic ingest unit test utility functions to toplevel ingest package module
